fixing

+++ core/product/src/product/product.service.ts
@@ -2,11 +2,15 @@ import { Injectable } from '@nestjs/common';
 import { BaseService } from 'src/interface/service/BaseService';
 import { Product } from './entity/product.entity';
 import { Repository } from 'typeorm';
+import { InjectRepository } from '@nestjs/typeorm';
 
 @Injectable()
 export class ProductService implements BaseService<Product> {
 
-  constructor(private readonly repo: Repository<Product>) { }
+  constructor(
+    @InjectRepository(Product)
+    private readonly repo: Repository<Product>
+  ) { }
 
   public async findOne(id: string): Promise<Product> {
     const product = await this.repo.findOne({
@@ -14,7 +18,7 @@ export class ProductService implements BaseService<Product> {
     });
 
     if (!product) {
-      throw new Error(`Product with id ${id} not found`);
+      throw new Error("Product not found");
     }
 
     return product;
@@ -44,28 +48,36 @@ export class ProductService implements BaseService<Product> {
 
   public async create(data: any): Promise<Product> {
     const product = this.repo.create(data);
-    return this.repo.save(product[0]);
+    return await this.repo.save(product[0]);
   }
 
   public async update(data: any, id: string): Promise<Product> {
-    const product = this.repo.findOne({
+    const product = await this.repo.findOne({
       where: {
         id
       }
     });
 
+    if (!product) {
+      throw new Error("Product not found");
+    }
+
     Object.assign(product, data);
 
-    return await this.repo.save(product[0]);
+    return await this.repo.save(product);
   }
 
 
   public async delete(id: string): Promise<void> {
-    const product = this.repo.findOne({
+    const product = await this.repo.findOne({
       where: {
         id
       }
     })
+
+    if (!product) {
+      throw new Error("Product not found");
+    }
 
     await this.repo.remove(product);
   }
